9requirements.txt .txt
Flask==3.0.0
psycopg[binary]==3.1.18
cloudinary==1.38.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0